kakapo-upline: only execute when point is at end of leading whitespace
This removes the edge case when we are typing out some text in the
current line, and accidentally press BACKSPACE when typing out a word in
all caps. In such a case, we most certainly want to just press
BACKSPACE, not remove any blank lines above/below the current line.