workaround for C multiline-style comment paragraphs
This is a workaround for the Linux Kernel Coding Style, which uses
"<TAB(s)><SPACE><ASTERISK><SPACE>..." for its multiline (long-style)
comments. It uses a number of TABs followed by a SPACE, and thus for
Kakapo uses a mixture of TABs and SPACES for the leading whitespace,
which has been unacceptable so far. Kakapo now recognizes a clean subset
of such "mixed" lines, treating them as valid leading whitespace and
does not error out.