kakapo-backspace: improved intra-line behavior
We aggressively delete lots of space characters (moving point to
tab-width-divisible columns) when point is in the middle of a line. This
way, we can freely press TAB and BACKSPACE when point is inside a line
after some text, as they behave like the opposites of each other.
(Before, pressing BACKSPACE at the middle of a line always deleted just
one character.)