kakapo-all-ktab: fix usage of string-match
This made it so that kakapo-all-ktab always returned true! `string-match
""` always returns true because it searches for the empty regular
expression, which is always true at position 0. What we wanted to do
instead was check if the string was _empty_; this is now accomplished by
using `string=` instead of `string-match`.