new `kakapo-strict' variable and kakapo-if
If we set `kakapo-strict' to `t', we get the old behavior of not doing
anything upon some indentation-related error. If set to `nil', we carry
on and perform operations (such as backspace still doing backspace).

We had to perform some refactoring by using the new `kakapo-if' macro.
We use a macro because all arguments to regular functions are evaluated
when they are assigned as variable arguments to the function.  When we
pass a function `func' to a "function", we don't want it to be evaluated
by virtue of it being assigned as an argument!