backspace: fix error on near-empty buffers
The problem was that the various let-bindings in `kakapo-backspace'
perform some calculations, one of which involves calling
`buffer-substring-no-properties' (namely, the `deletion-substr'
binding). We would get an error on very small buffers (e.g., *scratch*
buffer with just one or two characters in it) because the substring
calculation was incorrect --- we would end up assigning
`distance-to-prev-tab-width' to be the full tab-width (`8' for *scratch*
buffer) without checking the position of `point'.

We now make sure that we return the full `tab-width' value for
`distance-to-prev-tab-width' if point is not at the beginning of the
line, and also make sure to set it to `1' if we are using hard tabs.

Light testing shows that this gives the desired effect.