kakapo-open: for opening above, use current indentation
In the old behavior, opening a line above would search for a deeper level of
indentation just like opening a line below. But that led to an odd edge case
like the following:

	f = do
		foo $ do
			a
			b
			c
		d

If we now open a line above from line `d', we would indent to where `c' is (as
it is a deeper level of indentation). This goes against how vanilla Vim
behaves (where the indentation is set relative to `d', not `c').

This patch matches the Vim behavior.

Supersedes GitHub PR #2 (@mat8913).