kakapo-open: fine-tune behavior for opening from a blank line
When opening from a blank line, use the new variable
`kakapo-open-blank-line-search-indentation' for guidance. By default, we search
for the nearest indentation level when opening a line below (but not above).
This corresponds to a value of '(nil t).

For those who want to mimic Vim's default behavior, set it to '(nil nil) (as Vim
always opens without any indentation when opening from a blank line, in either
direction) --- like this

	(setq kakapo-open-blank-line-search-indentation '(nil nil))

in ~/.emacs.

For those feeling more adventurous, set it to '(t t).

This is a continuation of Github PR #3.